+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix [#163](https://github.com/squint-cljs/cherry/issues/163): Add assert to macros
 - Fix [#165](https://github.com/squint-cljs/cherry/issues/165): Fix ClojureScript protocol dispatch
+- Fix [#169](https://github.com/squint-cljs/cherry/issues/169): `extend-type` on `Object`
 
 ## 0.4.32 (2025-10-21)